Target .NET Standard 2.0 exclusively
When a library targets .NET Standard, there is no need to multi target netcoreapp versions that implement  the targeted .NET Standard version. This only makes the NuGet package grow in size without any added benefit.

The reason why you would multi-target is to take advantage of APIs in .NET Core that are not available in .NET Standard. This is not the case for this project where all APIs used are compatible with .NET Standard 2.0 (including the Portable.BouncyCastle dependency).

Also, by targeting .NET Standard 2.0, this makes PgpCore compatible with .NET Framework 4.7.2 and 4.8.

The README has also been updated to reflect that change.