Update Utiltities.cs to support PGP 2.6.x compatibility
PgpLiteralDataGenerator has a constructor that takes a bool indicating whether to use old format.
If true, it generates literal data objects in the old format. /// This is important if you need compatibility with PGP 2.6.x.

Added bool value, useOldFormat, to all relevant utilities methods constructing PgpLiteralDataGenerator object. Some of methods that instantiate the PgpLiteralDataGenerator don't use the Open() method that can use the old format, so not adding the arg for those utility methods. Default value false to not break any current client usage of the Utilities class. useOldFormat passed to the PgpLiteralDataGenerator constructor.

Add property UseOldFormat, pass to WriteStreamToLiteralData

BouncyCastle's PgpLiteralDataGenerator has a constructor that takes a bool indicating whether or not to use "oldFormat". This can be set to true to support 2.6.x version of PGP. Including support for this in PgpCore by adding a bool property UseOldFormat  to PGP class defaulting to false. Encrypt methods passes UseOldFormat as an argument to Utilities.WriteStreamToLiteralData which wraps the call to the PgpLiteralDataGenerator constructor.